requirements.txt
aioconsole==0.8.2
aiosqlite==0.21.0
alembic==1.17.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
billiard==4.2.3
black==25.9.0
celery==5.5.3
certifi==2025.10.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
isort==7.0.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pydantic==2.12.1
pydantic-settings==2.12.0
pydantic_core==2.41.3
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.2.0
redis==7.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
